requirements.txt
meshtastic
flask
ollama
pyserial
python-dotenv